Add support for passing a name to logger
