+++ src/utilities/helpers.js
@@ -17,3 +17,25 @@ export const getCallee = (depth?: number = 3): string => {
     return d.reduce((acc, s, i, a) => (s === 'lib' || s === 'node_modules' ? a[i + 1] : acc), '').split(':').shift();
   }
 };
+
+/**
+ * Strips plugin name from mandatory prefixes and namespaces.
+ * @param {String} name  Plugin name.
+ */
+export const normalizeName = (name: string): string => {
+  let normalizedName: string = name;
+
+  // if (name.indexOf('core.') >= 0) {
+  //   return name.substr(name.indexOf('core.') + 5);
+  // }
+
+  if (normalizedName.indexOf('/') >= 0) {
+    normalizedName = normalizedName.substr(normalizedName.indexOf('/') + 1);
+  }
+
+  if (normalizedName.indexOf('briskhome-') >= 0) {
+    normalizedName = normalizedName.substr(normalizedName.indexOf('briskhome-') + 10);
+  }
+
+  return normalizedName;
+};
