+++ src/utilities/plugins.js
@@ -54,25 +54,3 @@ export const disablePlugin = (directory: string)
     }
     return false;
   }).length;
-
-/**
- * Strips plugin name from mandatory prefixes and namespaces.
- * @param {String} name  Plugin name.
- */
-export const normalizeName = (name: string): string => {
-  let normalizedName: string = name;
-
-  // if (name.indexOf('core.') >= 0) {
-  //   return name.substr(name.indexOf('core.') + 5);
-  // }
-
-  if (normalizedName.indexOf('/') >= 0) {
-    normalizedName = normalizedName.substr(normalizedName.indexOf('/') + 1);
-  }
-
-  if (normalizedName.indexOf('briskhome-') >= 0) {
-    normalizedName = normalizedName.substr(normalizedName.indexOf('briskhome-') + 10);
-  }
-
-  return normalizedName;
-};
